## Summary

Add comprehensive performance benchmarking infrastructure for monitoring_system to establish baseline metrics and enable continuous performance tracking.

- Implement Google Benchmark-based performance tests for core monitoring components
- Add CI/CD workflow for automated benchmark execution on PRs and commits
- Create baseline documentation framework for performance tracking and regression detection

## Changes

### Benchmarks Implemented

1. **Performance Profiler Benchmarks** (`metric_collection_bench.cpp`)
   - Single sample recording latency
   - Multiple samples recording performance
   - Metrics retrieval performance (single and all)
   - Concurrent sample recording (multi-threaded)
   - Scoped timer overhead measurement

2. **Event Bus Benchmarks** (`event_bus_bench.cpp`)
   - **Status**: Placeholder implementation
   - **Reason**: Event bus interface not exposed in public API
   - Will be implemented when API becomes available

3. **System Resource Collection Benchmarks** (`collector_overhead_bench.cpp`)
   - **Status**: Placeholder implementation
   - **Reason**: Collector interfaces not exposed in public API
   - Will be implemented when API becomes available

### Infrastructure

- **CMake Integration**: 
  - Build configuration for benchmarks with Google Benchmark dependency
  - Proper common_system dependency resolution
  - Support for both local and CI builds
  
- **CI/CD Workflow**: 
  - Automated benchmark execution on Ubuntu 22.04 and macOS 13
  - Multi-repository checkout (monitoring_system + common_system)
  - Benchmark result artifact upload
  - Baseline comparison framework (ready for future use)
  
- **Documentation**: 
  - Baseline metrics framework in `docs/BASELINE.md`
  - Comprehensive benchmark README in `benchmarks/README.md`

### Target Performance Metrics

| Category | Metric | Target | Acceptable | Status |
|----------|--------|--------|------------|--------|
| Performance Profiler | Sample recording | < 100ns | < 1μs | ✅ Implemented |
| Performance Profiler | Metrics retrieval | < 1μs | < 10μs | ✅ Implemented |
| Event Bus | Publication latency | < 10μs | < 100μs | ⏳ Pending API |
| Event Bus | Throughput | > 100k events/s | > 50k events/s | ⏳ Pending API |
| System Collector | Collection latency | < 1ms | < 10ms | ⏳ Pending API |
| System Collector | CPU overhead | < 0.5% | < 2% | ⏳ Pending API |

## Test Plan

- [x] Build benchmarks successfully on Ubuntu and macOS
- [x] Execute benchmark suites without errors
- [x] Verify CI workflow runs correctly
- [x] Fix common_system dependency resolution in CI
- [x] Fix CMakeLists.txt target and include paths
- [ ] Review benchmark results for baseline establishment
- [ ] Document baseline metrics in `docs/BASELINE.md`
- [ ] Implement event bus benchmarks when API is available
- [ ] Implement collector benchmarks when API is available

## CI Status

✅ All benchmark checks passing:
- Ubuntu 22.04 + Clang: SUCCESS
- macOS 13 + Clang: SUCCESS
- Benchmark Report: SUCCESS

## Known Limitations

1. Event bus benchmarks are placeholders - will be implemented once the event bus interface is exposed in the public API
2. Collector overhead benchmarks are placeholders - will be implemented once collector interfaces are exposed in the public API
3. Only performance profiler benchmarks are fully functional in this initial implementation

## Next Steps

1. Run benchmarks locally and document baseline results
2. Expose event_bus and collector interfaces in public API
3. Implement full benchmark suites for all components
4. Establish performance regression thresholds for CI